use json formatter for events
